move internal/pubsub -> pubsub
