+++ pubsub/pub_test.go
@@ -5,8 +5,8 @@ import (
 	"testing"
 	"time"
 
-	"gocloud.dev/internal/pubsub"
-	"gocloud.dev/internal/pubsub/driver"
+	"gocloud.dev/pubsub"
+	"gocloud.dev/pubsub/driver"
 )
 
 type funcTopic struct {
