+++ pubsub/rabbitpubsub/rabbit.go
@@ -24,8 +24,8 @@ import (
 	"time"
 
 	"github.com/streadway/amqp"
-	"gocloud.dev/internal/pubsub"
-	"gocloud.dev/internal/pubsub/driver"
+	"gocloud.dev/pubsub"
+	"gocloud.dev/pubsub/driver"
 )
 
 type topic struct {
